refactor: use common sortQuestions
